# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EleevateOverseas Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /programs:
    get:
      operationId: listPrograms
      tags: [programs]
      summary: List programs
      parameters:
        - name: universityId
          in: query
          schema:
            type: string
        - name: degree
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Programs list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramListResponse"
    post:
      operationId: createProgram
      tags: [programs]
      summary: Create program (admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramBody"
      responses:
        "201":
          description: Created program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
  /programs/{id}:
    get:
      operationId: getProgram
      tags: [programs]
      summary: Get program by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
    patch:
      operationId: updateProgram
      tags: [programs]
      summary: Update program (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgramBody"
      responses:
        "200":
          description: Updated program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
    delete:
      operationId: deleteProgram
      tags: [programs]
      summary: Delete program (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Countries ──────────────────────────────────────────────────────────────
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List active job listings
      security: [{ bearerAuth: [] }]
      parameters:
        - name: type
          in: query
          schema: { type: string }
        - name: country
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Job listings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/JobListing"
                  total:
                    type: integer
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job listing
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                company: { type: string }
                description: { type: string }
                location: { type: string }
                country: { type: string }
                type: { type: string }
                salary: { type: string }
              required: [title, company, description]
      responses:
        "201":
          description: Created job listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListing"
  /my-job-applications:
    get:
      operationId: getMyJobApplications
      tags: [jobs]
      summary: Get current user's job applications
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobApplication"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a single job listing by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListing"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job listing (owner or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated job listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListing"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job listing (owner or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /jobs/{id}/apply:
    post:
      operationId: applyToJob
      tags: [jobs]
      summary: Apply to a job listing
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                coverLetter: { type: string }
                resumeUrl: { type: string }
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApplication"
  /jobs/{id}/applications:
    get:
      operationId: getJobApplications
      tags: [jobs]
      summary: Get applications for a job (listing owner only)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobApplication"
  /jobs/{id}/applications/{appId}:
    patch:
      operationId: updateJobApplicationStatus
      tags: [jobs]
      summary: Update application status (job listing owner only)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: appId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [applied, reviewing, shortlisted, rejected, hired]
              required: [status]
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApplication"

  # ── Careers ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    University:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        country:
          type: string
        city:
          type: string
        ranking:
          type: integer
        logoUrl:
          type: string
        imageUrl:
          type: string
        website:
          type: string
        description:
          type: string
        acceptanceRate:
          type: number
        avgTuitionUsd:
          type: integer
        featured:
          type: boolean
        programCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, country, city, createdAt]
...
    Program:
      type: object
      properties:
        id:
          type: string
        universityId:
          type: string
        name:
          type: string
        degree:
          type: string
          enum: [bachelor, master, phd, diploma, certificate, mba]
        field:
          type: string
        duration:
          type: integer
        durationUnit:
          type: string
          enum: [months, years]
        tuitionUsd:
          type: integer
        applicationDeadline:
          type: string
          format: date
        startDate:
          type: string
          format: date
        ieltsRequirement:
          type: number
        toeflRequirement:
          type: number
        gpaRequirement:
          type: number
        description:
          type: string
        university:
          $ref: "#/components/schemas/University"
        createdAt:
          type: string
          format: date-time
      required: [id, universityId, name, degree, field, createdAt]
    ProgramListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Program"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
...
    CreateProgramBody:
      type: object
      properties:
        universityId:
          type: string
        name:
          type: string
        degree:
          type: string
          enum: [bachelor, master, phd, diploma, certificate, mba]
        field:
          type: string
        duration:
          type: integer
        durationUnit:
          type: string
          enum: [months, years]
        tuitionUsd:
          type: integer
        applicationDeadline:
          type: string
          format: date
        startDate:
          type: string
          format: date
        ieltsRequirement:
          type: number
        toeflRequirement:
          type: number
        gpaRequirement:
          type: number
        description:
          type: string
      required: [universityId, name, degree, field]
    UpdateProgramBody:
      type: object
      properties:
        name:
          type: string
        degree:
          type: string
          enum: [bachelor, master, phd, diploma, certificate, mba]
        field:
          type: string
        duration:
          type: integer
        durationUnit:
          type: string
          enum: [months, years]
        tuitionUsd:
          type: integer
        applicationDeadline:
          type: string
          format: date
        startDate:
          type: string
          format: date
        ieltsRequirement:
          type: number
        toeflRequirement:
          type: number
        gpaRequirement:
          type: number
        description:
          type: string
...
    JobListing:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        company: { type: string }
        description: { type: string }
        location: { type: string, nullable: true }
        country: { type: string, nullable: true }
        type: { type: string, enum: [full-time, part-time, internship, contract] }
        salary: { type: string, nullable: true }
        skillsRequired: { type: array, items: { type: string } }
        status: { type: string, enum: [active, closed, draft] }
        logoUrl: { type: string, nullable: true }
        companyWebsite: { type: string, nullable: true }
        postedById: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, title, company, description, type, status, postedById, createdAt]
    JobApplication:
      type: object
      properties:
        id: { type: string }
        jobId: { type: string }
        userId: { type: string }
        coverLetter: { type: string, nullable: true }
        resumeUrl: { type: string, nullable: true }
        status: { type: string, enum: [applied, reviewing, shortlisted, rejected, hired] }
        createdAt: { type: string, format: date-time }
      required: [id, jobId, userId, status, createdAt]
...